Move ComponentRegister into helpers and break out LinkComponentType
