+++ docs/api/assets/js/search.js
@@ -1,3 +1,3 @@
 var typedoc = typedoc || {};
             typedoc.search = typedoc.search || {};
-            typedoc.search.data = {"kinds":{"1":"External module","32":"Variable","64":"Function","128":"Class","256":"Interface","1024":"Property","2048":"Method","2097152":"Object literal","4194304":"Type alias"},"rows":[{"id":0,"kind":1,"name":"\"ComponentRegister\"","url":"modules/_componentregister_.html","classes":"tsd-kind-external-module"},{"id":1,"kind":256,"name":"LinkComponentProps","url":"interfaces/_componentregister_.linkcomponentprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"ComponentRegister\""},{"id":2,"kind":1024,"name":"className","url":"interfaces/_componentregister_.linkcomponentprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"ComponentRegister\".LinkComponentProps"},{"id":3,"kind":1024,"name":"href","url":"interfaces/_componentregister_.linkcomponentprops.html#href","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"ComponentRegister\".LinkComponentProps"},{"id":4,"kind":1024,"name":"children","url":"interfaces/_componentregister_.linkcomponentprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"ComponentRegister\".LinkComponentProps"},{"id":5,"kind":256,"name":"ComponentRegisterComponents","url":"interfaces/_componentregister_.componentregistercomponents.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"ComponentRegister\""},{"id":6,"kind":1024,"name":"Link","url":"interfaces/_componentregister_.componentregistercomponents.html#link","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"ComponentRegister\".ComponentRegisterComponents"},{"id":7,"kind":256,"name":"ComponentRegisterOptions","url":"interfaces/_componentregister_.componentregisteroptions.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"ComponentRegister\""},{"id":8,"kind":1024,"name":"components","url":"interfaces/_componentregister_.componentregisteroptions.html#components","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"ComponentRegister\".ComponentRegisterOptions"},{"id":9,"kind":128,"name":"ComponentRegister","url":"classes/_componentregister_.componentregister.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"ComponentRegister\""},{"id":10,"kind":1024,"name":"Link","url":"classes/_componentregister_.componentregister.html#link","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"\"ComponentRegister\".ComponentRegister"},{"id":11,"kind":2048,"name":"reset","url":"classes/_componentregister_.componentregister.html#reset","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"ComponentRegister\".ComponentRegister"},{"id":12,"kind":2048,"name":"init","url":"classes/_componentregister_.componentregister.html#init","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"ComponentRegister\".ComponentRegister"},{"id":13,"kind":4194304,"name":"LinkComponent","url":"modules/_componentregister_.html#linkcomponent","classes":"tsd-kind-type-alias tsd-parent-kind-external-module","parent":"\"ComponentRegister\""},{"id":14,"kind":2097152,"name":"defaultComponentRegisterOptions","url":"modules/_componentregister_.html#defaultcomponentregisteroptions","classes":"tsd-kind-object-literal tsd-parent-kind-external-module","parent":"\"ComponentRegister\""},{"id":15,"kind":2097152,"name":"components","url":"modules/_componentregister_.html#defaultcomponentregisteroptions.components","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal","parent":"\"ComponentRegister\".defaultComponentRegisterOptions"},{"id":16,"kind":32,"name":"Link","url":"modules/_componentregister_.html#defaultcomponentregisteroptions.components.link","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"ComponentRegister\".defaultComponentRegisterOptions.components"},{"id":17,"kind":1,"name":"\"components/ErrorMessage\"","url":"modules/_components_errormessage_.html","classes":"tsd-kind-external-module"},{"id":18,"kind":256,"name":"ErrorMessageProps","url":"interfaces/_components_errormessage_.errormessageprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/ErrorMessage\""},{"id":19,"kind":1024,"name":"id","url":"interfaces/_components_errormessage_.errormessageprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/ErrorMessage\".ErrorMessageProps"},{"id":20,"kind":1024,"name":"className","url":"interfaces/_components_errormessage_.errormessageprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/ErrorMessage\".ErrorMessageProps"},{"id":21,"kind":1024,"name":"visuallyHiddenText","url":"interfaces/_components_errormessage_.errormessageprops.html#visuallyhiddentext","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/ErrorMessage\".ErrorMessageProps"},{"id":22,"kind":1024,"name":"children","url":"interfaces/_components_errormessage_.errormessageprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/ErrorMessage\".ErrorMessageProps"},{"id":23,"kind":64,"name":"ErrorMessage","url":"modules/_components_errormessage_.html#errormessage","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/ErrorMessage\""},{"id":24,"kind":1,"name":"\"components/Hint\"","url":"modules/_components_hint_.html","classes":"tsd-kind-external-module"},{"id":25,"kind":256,"name":"HintProps","url":"interfaces/_components_hint_.hintprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Hint\""},{"id":26,"kind":1024,"name":"id","url":"interfaces/_components_hint_.hintprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Hint\".HintProps"},{"id":27,"kind":1024,"name":"className","url":"interfaces/_components_hint_.hintprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Hint\".HintProps"},{"id":28,"kind":1024,"name":"children","url":"interfaces/_components_hint_.hintprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Hint\".HintProps"},{"id":29,"kind":64,"name":"Hint","url":"modules/_components_hint_.html#hint","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/Hint\""},{"id":30,"kind":1,"name":"\"components/Label\"","url":"modules/_components_label_.html","classes":"tsd-kind-external-module"},{"id":31,"kind":256,"name":"LabelProps","url":"interfaces/_components_label_.labelprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Label\""},{"id":32,"kind":1024,"name":"id","url":"interfaces/_components_label_.labelprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Label\".LabelProps"},{"id":33,"kind":1024,"name":"className","url":"interfaces/_components_label_.labelprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Label\".LabelProps"},{"id":34,"kind":1024,"name":"labelFor","url":"interfaces/_components_label_.labelprops.html#labelfor","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Label\".LabelProps"},{"id":35,"kind":1024,"name":"children","url":"interfaces/_components_label_.labelprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Label\".LabelProps"},{"id":36,"kind":64,"name":"Label","url":"modules/_components_label_.html#label","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/Label\""},{"id":37,"kind":1,"name":"\"components/Link\"","url":"modules/_components_link_.html","classes":"tsd-kind-external-module"},{"id":38,"kind":256,"name":"LinkProps","url":"interfaces/_components_link_.linkprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Link\""},{"id":39,"kind":1024,"name":"id","url":"interfaces/_components_link_.linkprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Link\".LinkProps"},{"id":40,"kind":1024,"name":"className","url":"interfaces/_components_link_.linkprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Link\".LinkProps"},{"id":41,"kind":1024,"name":"href","url":"interfaces/_components_link_.linkprops.html#href","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Link\".LinkProps"},{"id":42,"kind":1024,"name":"children","url":"interfaces/_components_link_.linkprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Link\".LinkProps"},{"id":43,"kind":64,"name":"Link","url":"modules/_components_link_.html#link","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/Link\""},{"id":44,"kind":1,"name":"\"components/PageAnnouncement\"","url":"modules/_components_pageannouncement_.html","classes":"tsd-kind-external-module"},{"id":45,"kind":256,"name":"PageAnnouncementProps","url":"interfaces/_components_pageannouncement_.pageannouncementprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/PageAnnouncement\""},{"id":46,"kind":1024,"name":"id","url":"interfaces/_components_pageannouncement_.pageannouncementprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/PageAnnouncement\".PageAnnouncementProps"},{"id":47,"kind":1024,"name":"className","url":"interfaces/_components_pageannouncement_.pageannouncementprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/PageAnnouncement\".PageAnnouncementProps"},{"id":48,"kind":1024,"name":"title","url":"interfaces/_components_pageannouncement_.pageannouncementprops.html#title","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/PageAnnouncement\".PageAnnouncementProps"},{"id":49,"kind":1024,"name":"children","url":"interfaces/_components_pageannouncement_.pageannouncementprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/PageAnnouncement\".PageAnnouncementProps"},{"id":50,"kind":64,"name":"PageAnnouncement","url":"modules/_components_pageannouncement_.html#pageannouncement","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/PageAnnouncement\""},{"id":51,"kind":1,"name":"\"components/Tag\"","url":"modules/_components_tag_.html","classes":"tsd-kind-external-module"},{"id":52,"kind":256,"name":"TagProps","url":"interfaces/_components_tag_.tagprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Tag\""},{"id":53,"kind":1024,"name":"id","url":"interfaces/_components_tag_.tagprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Tag\".TagProps"},{"id":54,"kind":1024,"name":"className","url":"interfaces/_components_tag_.tagprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Tag\".TagProps"},{"id":55,"kind":1024,"name":"children","url":"interfaces/_components_tag_.tagprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Tag\".TagProps"},{"id":56,"kind":64,"name":"Tag","url":"modules/_components_tag_.html#tag","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/Tag\""},{"id":57,"kind":1,"name":"\"components/Fieldset/Fieldset\"","url":"modules/_components_fieldset_fieldset_.html","classes":"tsd-kind-external-module"},{"id":58,"kind":256,"name":"FieldsetProps","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Fieldset/Fieldset\""},{"id":59,"kind":1024,"name":"id","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/Fieldset\".FieldsetProps"},{"id":60,"kind":1024,"name":"className","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/Fieldset\".FieldsetProps"},{"id":61,"kind":1024,"name":"role","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html#role","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/Fieldset\".FieldsetProps"},{"id":62,"kind":1024,"name":"legend","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html#legend","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/Fieldset\".FieldsetProps"},{"id":63,"kind":1024,"name":"describedBy","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html#describedby","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/Fieldset\".FieldsetProps"},{"id":64,"kind":1024,"name":"children","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/Fieldset\".FieldsetProps"},{"id":65,"kind":64,"name":"Fieldset","url":"modules/_components_fieldset_fieldset_.html#fieldset","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/Fieldset/Fieldset\""},{"id":66,"kind":1,"name":"\"components/Fieldset/FieldsetLegend\"","url":"modules/_components_fieldset_fieldsetlegend_.html","classes":"tsd-kind-external-module"},{"id":67,"kind":256,"name":"FieldsetLegendProps","url":"interfaces/_components_fieldset_fieldsetlegend_.fieldsetlegendprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Fieldset/FieldsetLegend\""},{"id":68,"kind":1024,"name":"id","url":"interfaces/_components_fieldset_fieldsetlegend_.fieldsetlegendprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/FieldsetLegend\".FieldsetLegendProps"},{"id":69,"kind":1024,"name":"className","url":"interfaces/_components_fieldset_fieldsetlegend_.fieldsetlegendprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/FieldsetLegend\".FieldsetLegendProps"},{"id":70,"kind":1024,"name":"isPageHeading","url":"interfaces/_components_fieldset_fieldsetlegend_.fieldsetlegendprops.html#ispageheading","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/FieldsetLegend\".FieldsetLegendProps"},{"id":71,"kind":1024,"name":"children","url":"interfaces/_components_fieldset_fieldsetlegend_.fieldsetlegendprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/FieldsetLegend\".FieldsetLegendProps"},{"id":72,"kind":64,"name":"FieldsetLegend","url":"modules/_components_fieldset_fieldsetlegend_.html#fieldsetlegend","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/Fieldset/FieldsetLegend\""}]};
+            typedoc.search.data = {"kinds":{"1":"External module","32":"Variable","64":"Function","128":"Class","256":"Interface","1024":"Property","2048":"Method","2097152":"Object literal","4194304":"Type alias"},"rows":[{"id":0,"kind":1,"name":"\"components/ErrorMessage\"","url":"modules/_components_errormessage_.html","classes":"tsd-kind-external-module"},{"id":1,"kind":256,"name":"ErrorMessageProps","url":"interfaces/_components_errormessage_.errormessageprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/ErrorMessage\""},{"id":2,"kind":1024,"name":"id","url":"interfaces/_components_errormessage_.errormessageprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/ErrorMessage\".ErrorMessageProps"},{"id":3,"kind":1024,"name":"className","url":"interfaces/_components_errormessage_.errormessageprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/ErrorMessage\".ErrorMessageProps"},{"id":4,"kind":1024,"name":"visuallyHiddenText","url":"interfaces/_components_errormessage_.errormessageprops.html#visuallyhiddentext","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/ErrorMessage\".ErrorMessageProps"},{"id":5,"kind":1024,"name":"children","url":"interfaces/_components_errormessage_.errormessageprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/ErrorMessage\".ErrorMessageProps"},{"id":6,"kind":64,"name":"ErrorMessage","url":"modules/_components_errormessage_.html#errormessage","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/ErrorMessage\""},{"id":7,"kind":1,"name":"\"components/Hint\"","url":"modules/_components_hint_.html","classes":"tsd-kind-external-module"},{"id":8,"kind":256,"name":"HintProps","url":"interfaces/_components_hint_.hintprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Hint\""},{"id":9,"kind":1024,"name":"id","url":"interfaces/_components_hint_.hintprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Hint\".HintProps"},{"id":10,"kind":1024,"name":"className","url":"interfaces/_components_hint_.hintprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Hint\".HintProps"},{"id":11,"kind":1024,"name":"children","url":"interfaces/_components_hint_.hintprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Hint\".HintProps"},{"id":12,"kind":64,"name":"Hint","url":"modules/_components_hint_.html#hint","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/Hint\""},{"id":13,"kind":1,"name":"\"components/Label\"","url":"modules/_components_label_.html","classes":"tsd-kind-external-module"},{"id":14,"kind":256,"name":"LabelProps","url":"interfaces/_components_label_.labelprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Label\""},{"id":15,"kind":1024,"name":"id","url":"interfaces/_components_label_.labelprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Label\".LabelProps"},{"id":16,"kind":1024,"name":"className","url":"interfaces/_components_label_.labelprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Label\".LabelProps"},{"id":17,"kind":1024,"name":"labelFor","url":"interfaces/_components_label_.labelprops.html#labelfor","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Label\".LabelProps"},{"id":18,"kind":1024,"name":"children","url":"interfaces/_components_label_.labelprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Label\".LabelProps"},{"id":19,"kind":64,"name":"Label","url":"modules/_components_label_.html#label","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/Label\""},{"id":20,"kind":1,"name":"\"helpers/LinkComponentType\"","url":"modules/_helpers_linkcomponenttype_.html","classes":"tsd-kind-external-module"},{"id":21,"kind":256,"name":"LinkComponentTypeProps","url":"interfaces/_helpers_linkcomponenttype_.linkcomponenttypeprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"helpers/LinkComponentType\""},{"id":22,"kind":1024,"name":"id","url":"interfaces/_helpers_linkcomponenttype_.linkcomponenttypeprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/LinkComponentType\".LinkComponentTypeProps"},{"id":23,"kind":1024,"name":"className","url":"interfaces/_helpers_linkcomponenttype_.linkcomponenttypeprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/LinkComponentType\".LinkComponentTypeProps"},{"id":24,"kind":1024,"name":"href","url":"interfaces/_helpers_linkcomponenttype_.linkcomponenttypeprops.html#href","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/LinkComponentType\".LinkComponentTypeProps"},{"id":25,"kind":1024,"name":"children","url":"interfaces/_helpers_linkcomponenttype_.linkcomponenttypeprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/LinkComponentType\".LinkComponentTypeProps"},{"id":26,"kind":4194304,"name":"LinkComponentType","url":"modules/_helpers_linkcomponenttype_.html#linkcomponenttype","classes":"tsd-kind-type-alias tsd-parent-kind-external-module","parent":"\"helpers/LinkComponentType\""},{"id":27,"kind":1,"name":"\"helpers/ComponentRegister\"","url":"modules/_helpers_componentregister_.html","classes":"tsd-kind-external-module"},{"id":28,"kind":256,"name":"ComponentRegisterComponents","url":"interfaces/_helpers_componentregister_.componentregistercomponents.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"helpers/ComponentRegister\""},{"id":29,"kind":1024,"name":"Link","url":"interfaces/_helpers_componentregister_.componentregistercomponents.html#link","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/ComponentRegister\".ComponentRegisterComponents"},{"id":30,"kind":256,"name":"ComponentRegisterOptions","url":"interfaces/_helpers_componentregister_.componentregisteroptions.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter","parent":"\"helpers/ComponentRegister\""},{"id":31,"kind":1024,"name":"components","url":"interfaces/_helpers_componentregister_.componentregisteroptions.html#components","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"helpers/ComponentRegister\".ComponentRegisterOptions"},{"id":32,"kind":128,"name":"ComponentRegister","url":"classes/_helpers_componentregister_.componentregister.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"helpers/ComponentRegister\""},{"id":33,"kind":1024,"name":"Link","url":"classes/_helpers_componentregister_.componentregister.html#link","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-static","parent":"\"helpers/ComponentRegister\".ComponentRegister"},{"id":34,"kind":2048,"name":"reset","url":"classes/_helpers_componentregister_.componentregister.html#reset","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"helpers/ComponentRegister\".ComponentRegister"},{"id":35,"kind":2048,"name":"init","url":"classes/_helpers_componentregister_.componentregister.html#init","classes":"tsd-kind-method tsd-parent-kind-class tsd-is-static","parent":"\"helpers/ComponentRegister\".ComponentRegister"},{"id":36,"kind":2097152,"name":"defaultComponentRegisterOptions","url":"modules/_helpers_componentregister_.html#defaultcomponentregisteroptions","classes":"tsd-kind-object-literal tsd-parent-kind-external-module","parent":"\"helpers/ComponentRegister\""},{"id":37,"kind":2097152,"name":"components","url":"modules/_helpers_componentregister_.html#defaultcomponentregisteroptions.components","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal","parent":"\"helpers/ComponentRegister\".defaultComponentRegisterOptions"},{"id":38,"kind":32,"name":"Link","url":"modules/_helpers_componentregister_.html#defaultcomponentregisteroptions.components.link","classes":"tsd-kind-variable tsd-parent-kind-object-literal","parent":"\"helpers/ComponentRegister\".defaultComponentRegisterOptions.components"},{"id":39,"kind":1,"name":"\"components/Link\"","url":"modules/_components_link_.html","classes":"tsd-kind-external-module"},{"id":40,"kind":256,"name":"LinkProps","url":"interfaces/_components_link_.linkprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Link\""},{"id":41,"kind":1024,"name":"id","url":"interfaces/_components_link_.linkprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Link\".LinkProps"},{"id":42,"kind":1024,"name":"className","url":"interfaces/_components_link_.linkprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Link\".LinkProps"},{"id":43,"kind":1024,"name":"href","url":"interfaces/_components_link_.linkprops.html#href","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Link\".LinkProps"},{"id":44,"kind":1024,"name":"children","url":"interfaces/_components_link_.linkprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Link\".LinkProps"},{"id":45,"kind":64,"name":"Link","url":"modules/_components_link_.html#link","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/Link\""},{"id":46,"kind":1,"name":"\"components/PageAnnouncement\"","url":"modules/_components_pageannouncement_.html","classes":"tsd-kind-external-module"},{"id":47,"kind":256,"name":"PageAnnouncementProps","url":"interfaces/_components_pageannouncement_.pageannouncementprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/PageAnnouncement\""},{"id":48,"kind":1024,"name":"id","url":"interfaces/_components_pageannouncement_.pageannouncementprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/PageAnnouncement\".PageAnnouncementProps"},{"id":49,"kind":1024,"name":"className","url":"interfaces/_components_pageannouncement_.pageannouncementprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/PageAnnouncement\".PageAnnouncementProps"},{"id":50,"kind":1024,"name":"title","url":"interfaces/_components_pageannouncement_.pageannouncementprops.html#title","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/PageAnnouncement\".PageAnnouncementProps"},{"id":51,"kind":1024,"name":"children","url":"interfaces/_components_pageannouncement_.pageannouncementprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/PageAnnouncement\".PageAnnouncementProps"},{"id":52,"kind":64,"name":"PageAnnouncement","url":"modules/_components_pageannouncement_.html#pageannouncement","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/PageAnnouncement\""},{"id":53,"kind":1,"name":"\"components/Tag\"","url":"modules/_components_tag_.html","classes":"tsd-kind-external-module"},{"id":54,"kind":256,"name":"TagProps","url":"interfaces/_components_tag_.tagprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Tag\""},{"id":55,"kind":1024,"name":"id","url":"interfaces/_components_tag_.tagprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Tag\".TagProps"},{"id":56,"kind":1024,"name":"className","url":"interfaces/_components_tag_.tagprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Tag\".TagProps"},{"id":57,"kind":1024,"name":"children","url":"interfaces/_components_tag_.tagprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Tag\".TagProps"},{"id":58,"kind":64,"name":"Tag","url":"modules/_components_tag_.html#tag","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/Tag\""},{"id":59,"kind":1,"name":"\"components/Fieldset/Fieldset\"","url":"modules/_components_fieldset_fieldset_.html","classes":"tsd-kind-external-module"},{"id":60,"kind":256,"name":"FieldsetProps","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Fieldset/Fieldset\""},{"id":61,"kind":1024,"name":"id","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/Fieldset\".FieldsetProps"},{"id":62,"kind":1024,"name":"className","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/Fieldset\".FieldsetProps"},{"id":63,"kind":1024,"name":"role","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html#role","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/Fieldset\".FieldsetProps"},{"id":64,"kind":1024,"name":"legend","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html#legend","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/Fieldset\".FieldsetProps"},{"id":65,"kind":1024,"name":"describedBy","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html#describedby","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/Fieldset\".FieldsetProps"},{"id":66,"kind":1024,"name":"children","url":"interfaces/_components_fieldset_fieldset_.fieldsetprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/Fieldset\".FieldsetProps"},{"id":67,"kind":64,"name":"Fieldset","url":"modules/_components_fieldset_fieldset_.html#fieldset","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/Fieldset/Fieldset\""},{"id":68,"kind":1,"name":"\"components/Fieldset/FieldsetLegend\"","url":"modules/_components_fieldset_fieldsetlegend_.html","classes":"tsd-kind-external-module"},{"id":69,"kind":256,"name":"FieldsetLegendProps","url":"interfaces/_components_fieldset_fieldsetlegend_.fieldsetlegendprops.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"components/Fieldset/FieldsetLegend\""},{"id":70,"kind":1024,"name":"id","url":"interfaces/_components_fieldset_fieldsetlegend_.fieldsetlegendprops.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/FieldsetLegend\".FieldsetLegendProps"},{"id":71,"kind":1024,"name":"className","url":"interfaces/_components_fieldset_fieldsetlegend_.fieldsetlegendprops.html#classname","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/FieldsetLegend\".FieldsetLegendProps"},{"id":72,"kind":1024,"name":"isPageHeading","url":"interfaces/_components_fieldset_fieldsetlegend_.fieldsetlegendprops.html#ispageheading","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/FieldsetLegend\".FieldsetLegendProps"},{"id":73,"kind":1024,"name":"children","url":"interfaces/_components_fieldset_fieldsetlegend_.fieldsetlegendprops.html#children","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"components/Fieldset/FieldsetLegend\".FieldsetLegendProps"},{"id":74,"kind":64,"name":"FieldsetLegend","url":"modules/_components_fieldset_fieldsetlegend_.html#fieldsetlegend","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"components/Fieldset/FieldsetLegend\""}]};
