+++ docs/api/globals.html
@@ -70,7 +70,6 @@
 						<section class="tsd-index-section ">
 							<h3>External modules</h3>
 							<ul class="tsd-index-list">
-								<li class="tsd-kind-external-module"><a href="modules/_componentregister_.html" class="tsd-kind-icon">&quot;<wbr>Component<wbr>Register&quot;</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_components_errormessage_.html" class="tsd-kind-icon">&quot;components/<wbr>Error<wbr>Message&quot;</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_components_fieldset_fieldset_.html" class="tsd-kind-icon">&quot;components/<wbr>Fieldset/<wbr>Fieldset&quot;</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_components_fieldset_fieldsetlegend_.html" class="tsd-kind-icon">&quot;components/<wbr>Fieldset/<wbr>Fieldset<wbr>Legend&quot;</a></li>
@@ -79,6 +78,8 @@
 								<li class="tsd-kind-external-module"><a href="modules/_components_link_.html" class="tsd-kind-icon">&quot;components/<wbr>Link&quot;</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_components_pageannouncement_.html" class="tsd-kind-icon">&quot;components/<wbr>Page<wbr>Announcement&quot;</a></li>
 								<li class="tsd-kind-external-module"><a href="modules/_components_tag_.html" class="tsd-kind-icon">&quot;components/<wbr>Tag&quot;</a></li>
+								<li class="tsd-kind-external-module"><a href="modules/_helpers_componentregister_.html" class="tsd-kind-icon">&quot;helpers/<wbr>Component<wbr>Register&quot;</a></li>
+								<li class="tsd-kind-external-module"><a href="modules/_helpers_linkcomponenttype_.html" class="tsd-kind-icon">&quot;helpers/<wbr>Link<wbr>Component<wbr>Type&quot;</a></li>
 							</ul>
 						</section>
 					</div>
@@ -90,9 +91,6 @@
 				<ul>
 					<li class="globals current ">
 						<a href="globals.html"><em>Globals</em></a>
-					</li>
-					<li class=" tsd-kind-external-module">
-						<a href="modules/_componentregister_.html">&quot;<wbr>Component<wbr>Register&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="modules/_components_errormessage_.html">&quot;components/<wbr>Error<wbr>Message&quot;</a>
@@ -117,6 +115,12 @@
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="modules/_components_tag_.html">&quot;components/<wbr>Tag&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="modules/_helpers_componentregister_.html">&quot;helpers/<wbr>Component<wbr>Register&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="modules/_helpers_linkcomponenttype_.html">&quot;helpers/<wbr>Link<wbr>Component<wbr>Type&quot;</a>
 					</li>
 				</ul>
 			</nav>
