+++ docs/api/interfaces/_components_pageannouncement_.pageannouncementprops.html
@@ -161,9 +161,6 @@
 						<a href="../globals.html"><em>Globals</em></a>
 					</li>
 					<li class=" tsd-kind-external-module">
-						<a href="../modules/_componentregister_.html">&quot;<wbr>Component<wbr>Register&quot;</a>
-					</li>
-					<li class=" tsd-kind-external-module">
 						<a href="../modules/_components_errormessage_.html">&quot;components/<wbr>Error<wbr>Message&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
@@ -186,6 +183,12 @@
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="../modules/_components_tag_.html">&quot;components/<wbr>Tag&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="../modules/_helpers_componentregister_.html">&quot;helpers/<wbr>Component<wbr>Register&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="../modules/_helpers_linkcomponenttype_.html">&quot;helpers/<wbr>Link<wbr>Component<wbr>Type&quot;</a>
 					</li>
 				</ul>
 			</nav>
