+++ docs/api/interfaces/_helpers_componentregister_.componentregisteroptions.html
@@ -56,10 +56,10 @@
 					<a href="../globals.html">Globals</a>
 				</li>
 				<li>
-					<a href="../modules/_componentregister_.html">&quot;ComponentRegister&quot;</a>
+					<a href="../modules/_helpers_componentregister_.html">&quot;helpers/ComponentRegister&quot;</a>
 				</li>
 				<li>
-					<a href="_componentregister_.componentregisteroptions.html">ComponentRegisterOptions</a>
+					<a href="_helpers_componentregister_.componentregisteroptions.html">ComponentRegisterOptions</a>
 				</li>
 			</ul>
 			<h1>Interface ComponentRegisterOptions&lt;C&gt;</h1>
@@ -72,19 +72,19 @@
 			<section class="tsd-panel tsd-comment">
 				<div class="tsd-comment tsd-typography">
 					<div class="lead">
-						<p>Options to pass to <a href="../classes/_componentregister_.componentregister.html#init">ComponentRegister.init</a>.</p>
+						<p>Options to pass to <a href="../classes/_helpers_componentregister_.componentregister.html#init">ComponentRegister.init</a>.</p>
 					</div>
 					<p>This is only generic to support sharing the type with
-						<a href="../modules/_componentregister_.html#defaultcomponentregisteroptions">defaultComponentRegisterOptions</a>. You should never need to override
+						<a href="../modules/_helpers_componentregister_.html#defaultcomponentregisteroptions">defaultComponentRegisterOptions</a>. You should never need to override
 						the default generic parameter. It defaults to
-					<a href="_componentregister_.componentregistercomponents.html">ComponentRegisterComponents</a>.</p>
+					<a href="_helpers_componentregister_.componentregistercomponents.html">ComponentRegisterComponents</a>.</p>
 				</div>
 			</section>
 			<section class="tsd-panel tsd-type-parameters">
 				<h3>Type parameters</h3>
 				<ul class="tsd-type-parameters">
 					<li>
-						<h4>C<span class="tsd-signature-symbol">: </span><a href="_componentregister_.componentregistercomponents.html" class="tsd-signature-type">ComponentRegisterComponents</a></h4>
+						<h4>C<span class="tsd-signature-symbol">: </span><a href="_helpers_componentregister_.componentregistercomponents.html" class="tsd-signature-type">ComponentRegisterComponents</a></h4>
 					</li>
 				</ul>
 			</section>
@@ -103,7 +103,7 @@
 						<section class="tsd-index-section ">
 							<h3>Properties</h3>
 							<ul class="tsd-index-list">
-								<li class="tsd-kind-property tsd-parent-kind-interface"><a href="_componentregister_.componentregisteroptions.html#components" class="tsd-kind-icon">components</a></li>
+								<li class="tsd-kind-property tsd-parent-kind-interface"><a href="_helpers_componentregister_.componentregisteroptions.html#components" class="tsd-kind-icon">components</a></li>
 							</ul>
 						</section>
 					</div>
@@ -117,14 +117,14 @@
 					<div class="tsd-signature tsd-kind-icon">components<span class="tsd-signature-symbol">:</span> <a href="" class="tsd-signature-type">C</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/LBHackney-IT/lbh-frontend-react/blob/master/src/ComponentRegister.ts#L48">ComponentRegister.ts:48</a></li>
+							<li>Defined in <a href="https://github.com/LBHackney-IT/lbh-frontend-react/blob/master/src/helpers/ComponentRegister.ts#L32">helpers/ComponentRegister.ts:32</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
 						<div class="lead">
 							<p>The components to register.</p>
 						</div>
-						<p>Defaults to be of type <a href="_componentregister_.componentregistercomponents.html">ComponentRegisterComponents</a>.</p>
+						<p>Defaults to be of type <a href="_helpers_componentregister_.componentregistercomponents.html">ComponentRegisterComponents</a>.</p>
 					</div>
 				</section>
 			</section>
@@ -134,9 +134,6 @@
 				<ul>
 					<li class="globals  ">
 						<a href="../globals.html"><em>Globals</em></a>
-					</li>
-					<li class="current tsd-kind-external-module">
-						<a href="../modules/_componentregister_.html">&quot;<wbr>Component<wbr>Register&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="../modules/_components_errormessage_.html">&quot;components/<wbr>Error<wbr>Message&quot;</a>
@@ -162,36 +159,36 @@
 					<li class=" tsd-kind-external-module">
 						<a href="../modules/_components_tag_.html">&quot;components/<wbr>Tag&quot;</a>
 					</li>
+					<li class="current tsd-kind-external-module">
+						<a href="../modules/_helpers_componentregister_.html">&quot;helpers/<wbr>Component<wbr>Register&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="../modules/_helpers_linkcomponenttype_.html">&quot;helpers/<wbr>Link<wbr>Component<wbr>Type&quot;</a>
+					</li>
 				</ul>
 			</nav>
 			<nav class="tsd-navigation secondary menu-sticky">
 				<ul class="before-current">
 					<li class=" tsd-kind-class tsd-parent-kind-external-module">
-						<a href="../classes/_componentregister_.componentregister.html" class="tsd-kind-icon">Component<wbr>Register</a>
+						<a href="../classes/_helpers_componentregister_.componentregister.html" class="tsd-kind-icon">Component<wbr>Register</a>
 					</li>
 					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
-						<a href="_componentregister_.componentregistercomponents.html" class="tsd-kind-icon">Component<wbr>Register<wbr>Components</a>
+						<a href="_helpers_componentregister_.componentregistercomponents.html" class="tsd-kind-icon">Component<wbr>Register<wbr>Components</a>
 					</li>
 				</ul>
 				<ul class="current">
 					<li class="current tsd-kind-interface tsd-parent-kind-external-module tsd-has-type-parameter">
-						<a href="_componentregister_.componentregisteroptions.html" class="tsd-kind-icon">Component<wbr>Register<wbr>Options</a>
+						<a href="_helpers_componentregister_.componentregisteroptions.html" class="tsd-kind-icon">Component<wbr>Register<wbr>Options</a>
 						<ul>
 							<li class=" tsd-kind-property tsd-parent-kind-interface">
-								<a href="_componentregister_.componentregisteroptions.html#components" class="tsd-kind-icon">components</a>
+								<a href="_helpers_componentregister_.componentregisteroptions.html#components" class="tsd-kind-icon">components</a>
 							</li>
 						</ul>
 					</li>
 				</ul>
 				<ul class="after-current">
-					<li class=" tsd-kind-interface tsd-parent-kind-external-module">
-						<a href="_componentregister_.linkcomponentprops.html" class="tsd-kind-icon">Link<wbr>Component<wbr>Props</a>
-					</li>
-					<li class=" tsd-kind-type-alias tsd-parent-kind-external-module">
-						<a href="../modules/_componentregister_.html#linkcomponent" class="tsd-kind-icon">Link<wbr>Component</a>
-					</li>
 					<li class=" tsd-kind-object-literal tsd-parent-kind-external-module">
-						<a href="../modules/_componentregister_.html#defaultcomponentregisteroptions" class="tsd-kind-icon">default<wbr>Component<wbr>Register<wbr>Options</a>
+						<a href="../modules/_helpers_componentregister_.html#defaultcomponentregisteroptions" class="tsd-kind-icon">default<wbr>Component<wbr>Register<wbr>Options</a>
 					</li>
 				</ul>
 			</nav>
