+++ docs/api/modules/_components_hint_.html
@@ -124,9 +124,6 @@
 						<a href="../globals.html"><em>Globals</em></a>
 					</li>
 					<li class=" tsd-kind-external-module">
-						<a href="_componentregister_.html">&quot;<wbr>Component<wbr>Register&quot;</a>
-					</li>
-					<li class=" tsd-kind-external-module">
 						<a href="_components_errormessage_.html">&quot;components/<wbr>Error<wbr>Message&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
@@ -149,6 +146,12 @@
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="_components_tag_.html">&quot;components/<wbr>Tag&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="_helpers_componentregister_.html">&quot;helpers/<wbr>Component<wbr>Register&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="_helpers_linkcomponenttype_.html">&quot;helpers/<wbr>Link<wbr>Component<wbr>Type&quot;</a>
 					</li>
 				</ul>
 			</nav>
