+++ docs/api/modules/_components_link_.html
@@ -104,7 +104,7 @@
 								<div class="lead">
 									<p>A component for creating an anchor linking to other pages. You can customize
 										the component used internally for integration with routers using
-									<a href="../classes/_componentregister_.componentregister.html#init">ComponentRegister.init</a>. By default, it uses an <code>&lt;a&gt;</code> tag.</p>
+									<a href="../classes/_helpers_componentregister_.componentregister.html#init">ComponentRegister.init</a>. By default, it uses an <code>&lt;a&gt;</code> tag.</p>
 								</div>
 							</div>
 							<h4 class="tsd-parameters-title">Parameters</h4>
@@ -124,9 +124,6 @@
 				<ul>
 					<li class="globals  ">
 						<a href="../globals.html"><em>Globals</em></a>
-					</li>
-					<li class=" tsd-kind-external-module">
-						<a href="_componentregister_.html">&quot;<wbr>Component<wbr>Register&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="_components_errormessage_.html">&quot;components/<wbr>Error<wbr>Message&quot;</a>
@@ -151,6 +148,12 @@
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="_components_tag_.html">&quot;components/<wbr>Tag&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="_helpers_componentregister_.html">&quot;helpers/<wbr>Component<wbr>Register&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="_helpers_linkcomponenttype_.html">&quot;helpers/<wbr>Link<wbr>Component<wbr>Type&quot;</a>
 					</li>
 				</ul>
 			</nav>
