+++ jest.setup.ts
@@ -1,7 +1,7 @@
 import Adapter from "enzyme-adapter-react-16";
 import { configure } from "enzyme";
 
-import { ComponentRegister } from "./src/ComponentRegister";
+import { ComponentRegister } from "./src/helpers/ComponentRegister";
 
 configure({ adapter: new Adapter() });
